# Should only check frozen fix #3737