before ( ce6c05600af11f540302145ff4b6988a5a37518a )

```
$ time ./build/host/bin/mruby benchmark/bm_so_lists.rb
./build/host/bin/mruby benchmark/bm_so_lists.rb  52.75s user 0.75s system 97% cpu 54.662 total
```

after

```
$ time ./build/host/bin/mruby benchmark/bm_so_lists.rb
./build/host/bin/mruby benchmark/bm_so_lists.rb  8.52s user 0.18s system 99% cpu 8.733 total
```